shutdownconfig: fix for ' in password
Looks weird but works.
Problem was that rc.shutdown is sourcing /tmp/shutdownconfig_results and if MYPASSWORD is e.g.: MYPASSWORD='a ' b' the script breaks of course.

Ok, the final step is 'read' in init:
oldIFS="$IFS"; IFS=
read -r -s MYPASS
IFS="$oldOFS"

But I wait with this until the init modified by Shino will be pushed first.
___________

Btw, I noticed that if I use some of (freshly reinstated) special chars while creating "light encrypted" savefile, it fails to mount at boot (in fact, for some reason, such a savefile isn't encrypted at all, it's empty, zeroes), but I'm not keen to waste time on searching for the cause, because this is a joke, not encryption, and IMO should be removed, as it gives users false sense of security.
I mean, c'mon - sharp 10 y.o. kid could break it in no time. ;)
Regular passwords like "My Password!" are still ok though.