wmexit, wmreboot, wmpoweroff, wmrestart: use busybox ps
In late 2013, when I firt experimented with building a puplet, I didn't know that packages-templates modified every package. So when everything was done, I placed some packages without the template filtering.

Everything seemed to be ok, but when I tried to shutdown, nothing happened.

ps was a symlink to busybox, but 'ps -C' required ps-FULL. 'ps' was a script, a hack to handle the variety of scripts that were already written.

To prevent such an incident from happening again, better use 'busybox ps'.