puppyapps: simplify command line parsing
This implements the same behavior as before and fixes possible issues in default*)

All switches seem to work as before, no problems so far.

default*)
I have tested this in command line, rox and pcmanfm and passes the test in all of them.

tested apps: defaultaudioplayer, defaultmediaplayer

The previous method worked randomly, some files with long names and/or unusual chars failed to open. In fact it wasn't working at all in pcmanfm.

Now any file or multiple files passed through puppyapps work perfectly and really works like a wrapper.

I added exec in the same script because some apps might expect the app to terminate when the job is done, making it a wrapper indeed.