edit-initramfs improvements
compr_func() has been rewritten because gunzip -t was returning 0
even if the file was corrupted or not gzipped... (gunzip 1.6)

I added cli functions to extract and recompress initrd.gz using
the directory it is located in as the working directory.

The old gui code is still there, I added code that uses the cli functions as well.

I tested gui/cli commands with an initrd.gz file with the
following compression methods:
* gz
* xz
* bzip2

I also added a command to remaster/update initrd.gz
directly from a woof-ce directory (there is no gui for this).