Enhance boot params, pupsfs, zdrv, fdrv, adrv, ydrv, psave
Basically merge the changes made in puppylinux-woof-CE/initrd_progs#8

It's for advanced usage, so users might not notice it anyway.

There are some other small changes and are documented in the
puppylinux-woof-CE/initrd_progs history.

This init doesn't have any particular change that requires some extra adjustments..
so it works the same with the curren setup.