xwin: logic to replace xorg hardware profiles
This attemps to replace the hardware profiles stuff
based on ddcprobe.

The  xorg.conf-auto-pc template is good enough..
it's pretty generic and very specific to puppy,
so it can work with a variety of
hardware from a 1997 video card to a 2016 one.

This new logic will attemp to run X a second time
if it fails the first time
 (falling back to a generic template..)

If the system does not crash after that
then there's a chance to run xorgwizard...