new probedisk, backup the old one
this new probedisk actually is a bit old now.

seems to work ok all the time, it's based in the old probedisk
but using a new approach as the one found in fatdog.

the way to know and fix possible issues is when somebody with very exotic
hardware is testing this and reports the issues to me..