add alsa init stuff from ubuntu yakkety
when there is a problem.. people at the forum will always
ask for this stuff first.. so I think it might be of use
in some obscure cases.

ubuntu has also added some config stuff, so it's worth
adding it here.

these files are stripped when then alsa-utils package
is processed... i think it's ok as long as the files
here are updated every once in a while (every two years or so)...