Cross-build fix
When doing a cross-build of slacko64-14.2 (target x86-64) using slacko-14.1 (host x86) I notice 1000 errors on this line because "Puppy Standard" directory has a space in it: http://oi65.tinypic.com/351x45d.jpg  (line 967 rationalise = line 1195 testing) and most icons in puppy do not display as consequence. Adding "" around it fixes it, and cross-build is a success! :)