init: tmpfs = personal storage..?
the tmpfs and aufs are mounted early as far as i can tell.
in the old init aufs is mounted as the last step.

so basically stuff happens.

so in the end: remount the tmpfs and increase the size of the
tmpfs filesystem according to the old rules...
but also add the total size of the sfs's loaded in ram...
to (compensate and) emulate the old behavior.

this produces a more "dynamic" personal storage that can have
a different size on each boot.. taking into account
sfs's loaded in ram that are now counted as "used space"