3builddistro-Z: remove xorg drivers-alternate stuff
this code relies on xorgwizard, but the current xorgwizard
does not handles this. and i don't plan to reimplement this.

the default 'bad' drivers were: nv, modesetting

besides the minimun xorg server version should be 1.11 (2011),
which works pretty well everywhere.

a replacement can be added as an option in _00build.conf
to remove specific xorg drivers.. a couple lines of code..
just let me know