Pupsave Backup 1.0
for testing

At this point this PHB is basically a complete rewrite.
I decided to change the name to Pupsave Backup

It adds support for savefolders and pupmode 13.

It defaults to write "snapshots" in the current pupsave parent dir

The way this script works makes it a candidate for Boot Manager...