bootflash: misc edits
this script does not seem to work (makebootflash), but it has a nice gui.

this commit is to basically make it more readable.

also use the new probedisk.. this script always accepted
a param as the drv... I accidentally wiped a hard disk.
Now the script always checks it's a usbflash before continuing.