Puppy 'applications.menu' for pcmanfm (and other file managers)
Just install pcmanfm with the ppm. Select a file and click on 'open with'..
you'll have a working nice dialog where you can easily associate apps to specific file types.

The source for this is lxde_puppy_menu-1.pet (@peabee might want to get rid of this)
There's also a couple more files in lxpup_config_sc-1702.pet

I had issues with the icons, so I used the official woofce icons in /usr/share/pixmaps/puppy
(i edited all the .directory files in /usr/share/desktop-directories.
to change the icon name/path and remove Comment= to make the files smaller)
Some other trivial edits (symlinks)

This can be improved, but so far it's working fine in dpup stretch within
the rox desktop with pcmanfm already as a built-in pkg.