Add overlayfs support to raspbian/stretch builds only
Raspbian kernels do not have aufs, so in order to experiment with
native arm builds on my pi3 I had to add support for overlayfs to
3builddistro-Z

This is implemented as a patch in raspbian/stretch/patches
so it will not affect other builds.  If there is interest in
implementing this directly in woof-CE to allow builds using
run_woof on systems lacking aufs, 2createpackages will need
overlayfs support for slackware builds.