remove references to wakepup2
wakepup2 is a 1.44mb floppy img with dos stuff

if you need wakepup2 to boot puppy, you're probably still living in 2003
if a cd/hd with a bootloader is not enough, then the hardware should rest is in pieces.

floppy disks are relics nowadays. not easy to get, very easy to damage..