slacko 14.1: fix/revert dbus_slack
The "machine-id" part of this patch added on March 28, 2016 has never worked for me in 14.1:
https://github.com/puppylinux-woof-CE/woof-CE/commit/d23a3ed987320f8bde804c47530e1d2e2b6a1270#diff-44f9fa248552f0d6f9b254c93f0aa62b

It does seem to work for 14.2 though. In 14.1 it creates errors for many programs (firefox, hexchat, gnome-mplayer I think):
https://ibb.co/d0W3Lw

Putting that code in if statement so it doesn't run in 14.1 seems to be one way to fix it (behavior before patch).