kernel-kit: Add ability to download kernel source from git repo
USE_GIT_KERNEL can be set to URL of repo and
USE_GIT_KERNEL_CONFIG can optionally be set to the path of a
config file in the git repo to be used as DOTconfig.  This config
is tweaked a little in kernel-kit/funcs.sh configure_git_kernel()

kernel_is_plus_version=yes will probably need to be set if git
repo has any changes compared to the official kernel sources.