kernel-kit: Add ability to download cross compilers from git repo
USE_GIT_X86_TOOLS can be set to the URL of a git repo containing
a 32bit cross compiler tool-chain.

USE_GIT_X86_64_TOOLS can be set to the URL of a git repo
containing a 64bit cross compiler tool-chain.

The cross compiler is expected to be in a bin directory in the git
repo which will be added to PATH in kernel-kit/build.sh