eventmanager: remove "experimental" automatic unmounting of partitions
it's a resource hog and doesn't seem to be that useful.
this was mostly test in very old puppies with the 'defunct' issue

i have never seen that issue in precise+ and once a device is truly "locked"
you just can't unmount it

this is to simplify the rox desktop. nothing of value is lost.