init-bootmenu: bugfixes
WARNING: If you want to test init-bootmenu make sure you make
backups of your save file/folder.  If init is called with the
wrong combination of DISTRO_SPECS / save file/folder it will
trigger a 'upgrade', if you cancel the 'upgrade' and choose to run
in ram, init may delete everything in your save file/folder even
if you choose not to save the ram session.