init: remove cryptoloop support
crypto_luks is identified as a filesystem by blkid
therefore no matter the filename or extension, it will always be correctly detected

that is a problem with cryptoloop, you have to rely on filenames: *_crypt*

so if you a have luks encrypted savefile with *_crypt* in the filename, it will be detected as luks

but if it's not a luks savefile, the init script will refuse to load it,
it will display a warning msg, telling you that cryptoloop is not supported
and that you can open the savefile by clicking on it (filemnt).

it will then continue booting without savefile.