init: [upgrades] fix 'savefile masacre' bug (PM 12->5) (#1335)
Unmount $SAVE_LAYER (pup_rw (12) or pup_ro1 (13))

savefile PM 12: /pup_rw is a mountpoint
savefile PM 13: /pup_rw is a symlink, /pup_ro1 is a mountpoint

This issue doesn't seem to happen with savefolders.

Once a savefile is rejected (or ignored) it must be successfully unmounted,
because PUPMODE changes from 12 to 5.. there are hacks at the end of the script
that empty pup_rw for after switching root in PUPMODE 5.

I'm probably adding code to deal with this situation for the first time
so probably proper handling of rejected or ignored savefiles
has been broken for a while.

Closes #1335