pup_event: let pmount deal with floppy disks/drives
eventmanager now says:

Linux does not support hotplug detection when a floppy disk is inserted or removed.
Please run pmount, if /sys/block/fd0 exists, it will be in the list.
However the floppy disk might not be inserted, or even the actual
floppy drive might not exist.. the BIOS is to blame.