remove "fallback if binary is missing" message
When I opened gcolor2 in the menu, I saw in the title "fallback if binary is missing". I think this might confuse the user and make them think "oh no, my binary is missing and I need to find it again" when the script is now the main gcolor2. The script seems to be working just as good as the original app so I think this message in brackets can be removed, or at least changed to something other than what it was before. Do you agree? :D