init: don't list compressed pupsaves [standard]
well... init was supposed to list tarballs created with pupsave backup
and that hasn't been working for more than a year

commit 8364f033bb602a7125f68e9811312540d5dc8128 "fixed" that issue

But i remember people complaining when the pupsave backups were in the pupsave list
since then some changes broke the code to list pupsave tarballs

And i think there are other ways to do this, sfs files?

If the pupsave backups are not listed, the code is useless and should be deleted

I was the one who added this code, and if someone
needs it for some reason let me know.