init: remove support for the underdog distro (should be a container/chroot)
Looking at the changes in init and rc.shutdown, so the underdog can be loaded and unloaded.

It seems to me that the underdog stuff is some sort of primitive chroot/container stuff,
a very quirky and dangerous way to achieve what others have been doing with chroot/containers for ages.

Underdog is just too invasive and can easily lead to a broken system, it requires more improvements and stuff.

If you want to use the package manager or apps of some other installed distro,
you normally do it through a chroot, chroot jail, container,
something like that, don't know much about it.

Basically you mount stuff, mount --bind or something,
maybe backup some files in the target system,
copy /etc/resolv.conf and maybe a few other files,
and that's it you're ready to run apps.
Then unmount stuff and restore a few files after the chroot ends..

Let's say there is already 1 underdog: the compat distro. With the underdog, there are 2 underdogs.

I hope you remove underdog from your buildsys and create something based on chroot or something.

Underdog is some sort of chroot/container that is loaded in init as
the lower layer, in which you boot some other distro using a puppy,
and puppy takes over.

underdog is something huge, its implications are huge.
The tools that deal with remastering puppy will find trouble with a
loaded underdog and so on.

Answer to underdog: containers, chroot, etc. Problems: almost none.
Problems with underdog: many.

You can use any distro as the underdog and that distro may be highly
incompatible with the puppy you're using, i mean extremely
incompatible. Binary clashes, etc.

And aufs being just one of the possible unionfs, i guess underdog is
not properly supported by overlay or something, there are many things
that aufs can do that are not possible in overlay, with this a step to
unification is achieved..

The underdog is the product of old concepts and limitations of puppy.

With woof2 many of the limitations were overcome.. by using another
distro as the puppy itself.

With Pkg another limitation is being overcome, now it's possible to
use cli pet pkg manager.

With wine it's possible to use many applications, that if you look
carefully, it will satisfy basically all your needs, so there's no
need for something as tricky as this.

And the asnwer to everything: portable packages.

Packages compiled in a fairly old distro with some static stuff and
with very basic dependencies that will run basically on most newer
distros.

I think someone produced something like that: a fairly static mpv, it
was compiled in precise, i read people uses is the slackos and all
newer ubuntu puppies, even in wary, racy (with glibc upgrade).

You see, there's no need for this..