rpi: update package lists and repos for raspbian buster
This builds a reasonably sane distro that boots to X and runs mostly fine.
A modprobe of 'brcmfmac' is needed to get wireless working then you
can connect and browse with midori browser. It's the gtk3 version and
runs very smoothly.
There is lots to do to get this to alpha standard. Many old packages just
refuse to run, but they were built on Barry's old sap6 from years ago.
Some bloat needs cutting as well.