kernel-kit: add config for 5.4 LTS
I tried the config for EasyOS by BarryK (compiled with Debian) in Slacko64 and it seemed to work on a Core2Duo laptop and Cherry Trail Atom netbook. Here was the resulting config it made. Free free to use it or change it or toss it and use something better if you think it's not good. I upload a kernel made with it here too if you want to test:

https://transfernow.net/93cbe8c02tkb
md5sum: 125579e5c44879c726f79ae97e946d37

It is large because it comes with all firmware option and compressed with just -comp xz. I thought it mounted drives a bit slower when clicking on their icon and firewall gave me a warning about cdrecord config but those were probably unrelated.