Create xorg-autoconf
xorg-autoconf is a utility to generate a full xorg configuration with high accuracy and precision. It properly configures the graphics cards with proper drivers and settings depending on the video card installed on the computer. It gives also finer control for the user on some graphics cards (if a user has an experience on bash scripting).