installpkg.sh: Fixed package entry (CRITICAL)
Sometimes petget was unable to parse filename correctly that will lead to creation of this file "/var/packages/.files" which caused by DLPKG_NAME variable returning empty string if the specific package was not found in the repo database. This creates blank entry or malformed entry where first and second column of package entry was missing on /var/packages/user-installed-package

This will cause serious trouble because the user was unable to remove package and petget cant pinpoint the list of files to be deleted

This commit will solved this issue.