rc.shutdown: Add answer file for PUPMODE 13
The answer file for PUPMODE 13 if auto saved ramdisk session was disabled was extremely handy for shutdown scenario such as calling shutdown/reboot from ConsoleKit. It works like shutdownconfig. While the user is logged in and before it can ask user whether save the ramdisk and save the user response before start the shutdown process rather asking the user while in the middle of shutdown process