Updated new2dir
Changes:

1. don't include cmake stuff inside the built PETs (in `/root/<stuff that shouldn't be there>`)
2. find package version from git repo info, if possible

Change 1 is a fix, to avoid bundling build info stuff into the final contents 
of the PETs generated.

Change 2 is a new feature - get program version from date of build and 
git repo info. This  allows users to run `new2dir` in a folder that is a git
repo, even if it doesn't have a version in the parent folder name.