rc.sysinit: Added isolated folder
Due to unorthodox puppy scripts which has filenames having conflicts with the mainstream distro package. It can cause troubles when it was overwritten. The solution here was to  puppy scripts on a isolate folder and set it as first priority on $PATH this will prevent puppy scripts from being overwritten