pMusic: Move Streamripper from Recommended to Optional dependency
Streamripper is rather outdated, and many streams aren't working with it. The benefit of playing radio streams through Streamripper is to grab metainfo in real-time. In the future I guess we will be able to do this in ffmpeg itself by setting up a stream-server, but there is not much ducumentation around this topic...