Update init: use binded mount rather than direct path
Reason: If savefolder was used and the actual path of savefolder was used. It would not release from layered root filesystem and unmount the drive when the save folder was located during shutdown. This may cause filesystem corruption. In order to detach the savefolder gracefully from layer root filesystem in aufs was to use bind mountpoint. In order to retain the attached path on aufs layered system.